Handle the function name like '__lambda_[0-9]*__' in emitter instead of astdump.sh

+++ src/MapleFE/test/astdump.sh
@@ -14,7 +14,7 @@ Short/long options:
   -k | --keep            Keep generated files *.ts-[0-9]*.out.ts which fail to compile with tsc
   -A | --all             Process all .ts files in current directory excluding *.ts-[0-9]*.out.ts
   -C | --clean           Clean up generated files (*.ts-[0-9]*.out.ts)
-  -n | --name            Keep original names by removing "__lambda_[0-9]*__" and "__v[0-9]*" from generated code
+  -n | --name            Keep original names by removing "__v[0-9]*" from generated code
   -t | --treediff        Compare the AST of generated TS code with the one of original TS code
   -T | --Treediff        Same as -t/--treediff except that it disables tsc for generated TS code
   <file1> [<file2> ...]  Specify one or more TypeScript files to be processed
@@ -97,7 +97,7 @@ for ts in $LIST; do
   if [ "x${cmd:0:4}" = "xsed " ]; then
     T=$ts-$PROCID.out.ts
     eval $cmd <<< "$out" > "$T"
-    [ -z "$NAME" ] || sed -i -e 's/__v[0-9][0-9]*//g' -e 's/ __lambda_[0-9][0-9]*__/ /' -e 's/function *\((.*) => \)/\1/' "$T"
+    [ -z "$NAME" ] || sed -i 's/__v[0-9][0-9]*//g' "$T"
     clang-format-10 -i --style="{ColumnLimit: 120}" "$T"
     echo -e "\n====== TS Reformatted ======\n"
     $HIGHLIGHT "$T"
@@ -116,7 +116,7 @@ for ts in $LIST; do
     elif [ -z $TREEDIFF ]; then
       echo "MSG: Passed, test case $ts"
       Passed="$Passed $ts"
-      rm -f "$T"
+      [ -n "$KEEP" ] || rm -f "$T"
     else
       clang-format-10 --style="{ColumnLimit: 120}" $ts > $ts.tmp.ts
       $TS2AST $ts.tmp.ts
